Start making changes to causal save/load
